Conflicts:
	DNN Platform/Library/Services/Scheduling/Scheduler.cs
	DNN Platform/Modules/CoreMessaging/CoreMessaging.dnn
	DNN Platform/Modules/DDRMenu/DDRMenu.dnn
	DNN Platform/Modules/DigitalAssets/dnn_DigitalAssets.dnn
	DNN Platform/Modules/Groups/SocialGroups.dnn
	DNN Platform/Modules/HTML/dnn_HTML.dnn
	DNN Platform/Modules/Journal/Journal.dnn
	DNN Platform/Modules/MemberDirectory/MemberDirectory.dnn
	DNN Platform/Modules/MobileManagement/dnn_MobileManagement.dnn
	DNN Platform/Modules/PreviewProfileManagement/dnn_PreviewProfileManagement.dnn
	DNN Platform/Modules/RazorHost/Manifest.dnn
	DNN Platform/Modules/RazorHost/RazorHost.dnn
	DNN Platform/Modules/Taxonomy/Taxonomy.dnn
	DNN Platform/Modules/UrlManagement/dnn_UrlManagement.dnn
	DNN Platform/Providers/AuthenticationProviders/DotNetNuke.Authentication.Facebook/Facebook_Auth.dnn
	DNN Platform/Providers/AuthenticationProviders/DotNetNuke.Authentication.Google/Google_Auth.dnn
	DNN Platform/Providers/AuthenticationProviders/DotNetNuke.Authentication.LiveConnect/Live_Auth.dnn
	DNN Platform/Providers/AuthenticationProviders/DotNetNuke.Authentication.Twitter/Twitter_Auth.dnn
	DNN Platform/Providers/ClientCapabilityProviders/Provider.FiftyOneClientCapabilityProvider/FiftyOneClientCapabilityProvider.dnn
	DNN Platform/Providers/HtmlEditorProviders/RadEditorProvider/DotNetNuke.RadEditorProvider.dnn
	DNN Platform/Providers/NavigationProviders/ASP2MenuNavigationProvider/ASP2MenuNavigationProvider.dnn
	DNN Platform/Providers/NavigationProviders/DNNDropDownNavigationProvider/DNNDropDownNavigationProvider.dnn
	DNN Platform/Providers/NavigationProviders/DNNMenuNavigationProvider/DNNMenuNavigationProvider.dnn
	DNN Platform/Providers/NavigationProviders/DNNTreeNavigationProvider/DNNTreeNavigationProvider.dnn
	DNN Platform/Providers/NavigationProviders/SolpartMenuNavigationProvider/SolpartMenuNavigationProvider.dnn
	SolutionInfo.cs
	Website/DesktopModules/Admin/Console/ViewConsole.ascx.cs
	Website/Providers/DataProviders/SqlDataProvider/07.03.00.SqlDataProvider
	Website/Providers/DataProviders/SqlDataProvider/07.03.01.SqlDataProvider
	Website/js/Debug/dnn.modalpopup.js
	Website/js/dnn.modalpopup.js

+++ DNN Platform/Modules/DigitalAssets/Components/Controllers/DigitalAssetsController.cs
@@ -20,6 +20,7 @@
 #endregion
 
 using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.ComponentModel.Composition;
@@ -41,6 +42,7 @@ using DotNetNuke.Entities.Portals;
 using DotNetNuke.Entities.Users;
 using DotNetNuke.Modules.DigitalAssets.Components.Controllers.Models;
 using DotNetNuke.Modules.DigitalAssets.Components.ExtensionPoint;
+using DotNetNuke.Modules.DigitalAssets.Services.Models;
 using DotNetNuke.Security.Permissions;
 using DotNetNuke.Security.Roles;
 using DotNetNuke.Services.FileSystem;
@@ -56,7 +58,7 @@ namespace DotNetNuke.Modules.DigitalAssets.Components.Controllers
     public class DigitalAssetsController : IDigitalAssetsController, IUpgradeable
     {
         protected static readonly DigitalAssetsSettingsRepository SettingsRepository = new DigitalAssetsSettingsRepository();
-
+        private static readonly Hashtable MappedPathsSupported = new Hashtable();
         #region Static Private Methods
         private static bool IsHostMenu
         {
@@ -332,6 +334,31 @@ namespace DotNetNuke.Modules.DigitalAssets.Components.Controllers
 
             var result = newPath.Substring(0, nameIndex) + newPath.Substring(nameIndex).Replace(folderName, newFolderName);
             return result;
+        }
+
+        private bool AreMappedPathsSupported(int folderMappingId)
+        {
+            if (MappedPathsSupported.ContainsKey(folderMappingId))
+            {
+                return (bool)MappedPathsSupported[folderMappingId];
+            }
+            var result = FolderProvider.Instance(FolderMappingController.Instance.GetFolderMapping(folderMappingId).FolderProviderType).SupportsMappedPaths;
+            MappedPathsSupported.Add(folderMappingId, result);
+            return result;
+        }
+
+        private string GetUnlinkAllowedStatus(IFolderInfo folder)
+        {
+            if (AreMappedPathsSupported(folder.FolderMappingID) && GetFolder(folder.ParentID).FolderMappingID != folder.FolderMappingID)
+            {
+                return "onlyUnlink";
+            }
+            if (AreMappedPathsSupported(folder.FolderMappingID))
+            {
+                return "true";
+            }
+            return "false";
+
         }
         #endregion
 
@@ -635,7 +662,7 @@ namespace DotNetNuke.Modules.DigitalAssets.Components.Controllers
             }
             catch (FolderAlreadyExistsException)
             {
-                throw new DotNetNukeException(LocalizationHelper.GetString("FolderAlreadyExists.Error"));
+                throw new DotNetNukeException(string.Format(LocalizationHelper.GetString("FolderAlreadyExists.Error"), filterFolderName));
             }
         }
 
@@ -644,7 +671,35 @@ namespace DotNetNuke.Modules.DigitalAssets.Components.Controllers
             return GetItemViewModel(FileManager.Instance.GetFile(fileID, true));
         }
 
-        public IEnumerable<ItemPathViewModel> DeleteItems(IEnumerable<ItemBaseViewModel> items)
+        public void UnlinkFolder(int folderID)
+        {
+            var folder = FolderManager.Instance.GetFolder(folderID);
+            // Check if user has appropiate permissions
+            if (!HasPermission(folder, "DELETE"))
+            {
+                throw new DotNetNukeException(LocalizationHelper.GetString("UserHasNoPermissionToUnlinkFolder.Error"));
+            }
+            FolderManager.Instance.UnlinkFolder(folder);
+        }
+
+        public int GetMappedSubFoldersCount(IEnumerable<ItemBaseViewModel> items, int portalID)
+        {
+            var totalSubfoldersCount = 0;
+            if (items.All(i => !i.IsFolder))
+            {
+                return totalSubfoldersCount;
+            }
+            var allFolders = FolderManager.Instance.GetFolders(portalID);
+            foreach (var item in items.Where(i => i.IsFolder && HasPermission(FolderManager.Instance.GetFolder(i.ItemID), "VIEW")))
+            {
+                var folder = FolderManager.Instance.GetFolder(item.ItemID);
+                var allSubFolders = allFolders.Where(f => f.FolderPath.StartsWith(folder.FolderPath));
+                totalSubfoldersCount = totalSubfoldersCount + allSubFolders.Count(f => GetUnlinkAllowedStatus(f) == "onlyUnlink");
+            }
+            return totalSubfoldersCount;
+        }
+
+        public IEnumerable<ItemPathViewModel> DeleteItems(IEnumerable<DeleteItem> items)
         {
             var notDeletedItems = new List<ItemPathViewModel>();
 
@@ -652,7 +707,7 @@ namespace DotNetNuke.Modules.DigitalAssets.Components.Controllers
             {
                 if (item.IsFolder)
                 {
-                    var folder = FolderManager.Instance.GetFolder(item.ItemID);
+                    var folder = FolderManager.Instance.GetFolder(item.ItemId);
                     if (folder == null) continue;
 
                     if (!HasPermission(folder, "DELETE"))
@@ -661,12 +716,19 @@ namespace DotNetNuke.Modules.DigitalAssets.Components.Controllers
                     }
                     else
                     {
-                        DeleteFolder(folder, notDeletedItems);
+                        if (item.UnlinkAllowedStatus == "onlyUnlink")
+                        {
+                            FolderManager.Instance.UnlinkFolder(folder);
+                        }
+                        else
+                        {
+                            DeleteFolder(folder, notDeletedItems);
+                        }
                     }
                 }
                 else
                 {
-                    var fileInfo = FileManager.Instance.GetFile(item.ItemID, true);
+                    var fileInfo = FileManager.Instance.GetFile(item.ItemId, true);
                     if (fileInfo == null) continue;
 
                     var folder = FolderManager.Instance.GetFolder(fileInfo.FolderId);
@@ -980,7 +1042,7 @@ namespace DotNetNuke.Modules.DigitalAssets.Components.Controllers
                 ? LocalizationHelper.GetString("RootFolder.Text")
                 : folder.FolderName;
 
-            return new FolderViewModel
+            var folderViewModel = new FolderViewModel
             {
                 FolderID = folder.FolderID,
                 FolderMappingID = folder.FolderMappingID,
@@ -992,6 +1054,8 @@ namespace DotNetNuke.Modules.DigitalAssets.Components.Controllers
                 Permissions = GetPermissionViewModelCollection(folder),
                 HasChildren = folder.HasChildren
             };
+            folderViewModel.Attributes.Add(new KeyValuePair<string, object>("UnlinkAllowedStatus", GetUnlinkAllowedStatus(folder)));
+            return folderViewModel;
         }
 
         protected virtual ItemViewModel GetItemViewModel(IFolderInfo folder)
@@ -1017,10 +1081,11 @@ namespace DotNetNuke.Modules.DigitalAssets.Components.Controllers
                 Permissions = GetPermissionViewModelCollection(folder),
                 ParentFolderID = parentFolderId,
                 ParentFolder = parentFolderPath,
-                FolderMappingID = folder.FolderMappingID
+                FolderMappingID = folder.FolderMappingID,
+                UnlinkAllowedStatus = GetUnlinkAllowedStatus(folder)
             };
         }
-
+        
         protected virtual ItemViewModel GetItemViewModel(IFileInfo file)
         {
             var folder = FolderManager.Instance.GetFolder(file.FolderId);
@@ -1035,7 +1100,8 @@ namespace DotNetNuke.Modules.DigitalAssets.Components.Controllers
                 Permissions = GetPermissionViewModelCollection(folder),
                 ParentFolderID = folder.FolderID,
                 ParentFolder = folder.FolderPath,
-                Size = string.Format(new FileSizeFormatProvider(), "{0:fs}", file.Size)
+                Size = string.Format(new FileSizeFormatProvider(), "{0:fs}", file.Size),
+                UnlinkAllowedStatus = "false"
             };
         }
